rename sae.py to sparse_coder.py
