+++ README.md
@@ -49,7 +49,7 @@ To train SAEs from the command line, you can use the following command:
 ```bash
 python -m sparsify EleutherAI/pythia-160m <optional dataset>
 ```
-By default, we use the `EleutherAI/fineweb-edu-dedup-10b` dataset for training, but you can use any dataset HuggingFace Hub
+By default, we use the `EleutherAI/fineweb-edu-dedup-10b` dataset for training, but you can use any dataset from the HuggingFace Hub, or any local dataset in HuggingFace format (the string is passed to `load_dataset` from the `datasets` library).
 
 The CLI supports all of the config options provided by the `TrainConfig` class. You can see them by running `python -m sparsify --help`.
